% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\name{fhi_pal}
\alias{fhi_pal}
\title{FHI Palette Generator}
\usage{
fhi_pal(palette = "qualitative", reverse = FALSE, ...)
}
\arguments{
\item{palette}{Character name of palette.}

\item{reverse}{Boolean indicating whether the palette should be reversed.}

\item{...}{Additional arguments to pass to colorRampPalette()}
}
\description{
FHI Palette Generator
}
