% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\docType{data}
\name{fhi_palettes}
\alias{fhi_palettes}
\title{FHI Color Palettes}
\format{
An object of class \code{list} of length 7.
}
\usage{
fhi_palettes
}
\description{
FHI Color Palettes
}
\keyword{datasets}
