% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\name{scale_color_fhi}
\alias{scale_color_fhi}
\title{Color scale for FHI palettes}
\usage{
scale_color_fhi(palette = "qualitative", discrete = TRUE, reverse = FALSE, ...)
}
\arguments{
\item{palette}{Character name of palette.}

\item{discrete}{Boolean indicating whether color aesthetic is discrete.}

\item{reverse}{Boolean indicating whether the palette should be reversed.}

\item{...}{Additional arguments passed to discrete_scale() or scale_color_gradientn()}
}
\description{
Color scale for FHI palettes
}
