% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\name{scale_fill_fhi}
\alias{scale_fill_fhi}
\title{Fill scale for FHI palettes}
\usage{
scale_fill_fhi(palette = "qualitative", discrete = TRUE, reverse = FALSE, ...)
}
\arguments{
\item{palette}{Character name of palette.}

\item{discrete}{Boolean indicating whether fill aesthetic is discrete.}

\item{reverse}{Boolean indicating whether the palette should be reversed.}

\item{...}{Additional arguments passed to discrete_scale() or scale_fill_gradientn()}
}
\description{
Fill scale for FHI palettes
}
