% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\docType{data}
\name{fhi_palettes}
\alias{fhi_palettes}
\title{FHI Color Palettes}
\format{
An object of class \code{list} of length 9.
}
\usage{
fhi_palettes
}
\description{
Sequential ramps \code{grable} and \code{greige} are gråblå and greige from the FHI
visual profile (see \url{https://www.fhi.no/om/fhi/visuell-profil/innhold/farger/}),
ordered light to dark like the other quantitative palettes.
}
\keyword{datasets}
